Backend - Sprint 2 Updates



Lots of changes to the backend....

Authorization
Now almost all routes require some form of auth token to be sent. When 'signing' in a JWT token is sent in the response. This needs to be included in most routes to avoid a 403 - Forbidden response.

Login and Register (and a few test cases) also do not require JWT bearer tokens

You can now only get data for your own role. Example: User1 cannot get transactions for User2 and vis versa

Admin accounts can supersede some/all rules of authorization. Currently admin accounts cannot be created but are supported and one exists by default.

Additionally, the login API page has been updated to store the token in localstorage (this is temporary) and will include it when loading the home page. You can use this as an example of how-to

+++ asp.net_server/Controllers/AccountsController.cs
@@ -1,9 +1,14 @@
 using App.Models;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.EntityFrameworkCore;
+using System.Security.Claims;
+using System.Threading.Tasks;
+using System.Linq.Expressions;
 
 namespace App.Controllers;
 
+[Authorize]
 [ApiController]
 [Route("api/[controller]")]
 
@@ -16,12 +21,15 @@ public class AccountsController : ControllerBase
         _context = context;
     }
 
+    [AllowAnonymous]
     [HttpGet()]
-    public bool check()
-    {
-        return true;
-    }
+    public bool check() { return true; }
+
+    [Authorize]
+    [HttpGet("auth")]
+    public bool checkAuth() { return true; }
     
+    [Authorize(Roles = "Admin")]
     [HttpGet("GetAll")]
     public async Task<ActionResult<IEnumerable<Account>>> GetAccounts()
     {
@@ -34,6 +42,8 @@ public class AccountsController : ControllerBase
     [HttpGet("{Id}")]
     public async Task<ActionResult<Account>> GetAccount(int Id)
     {
+        if (!await AuthorizeUser(Id)) return Forbid();
+
         var account = await _context.Accounts
             // .Include(a => a.Transactions)
             // .Include(a => a.UserAccounts)
@@ -49,10 +59,7 @@ public class AccountsController : ControllerBase
 
     [HttpPost()]
     public async Task<ActionResult<Account>> PostAccount(Account account)
-    {
-
-        var a = await _context.Accounts.FindAsync(account.Id);
-        if (a != null) return BadRequest($"Account already exists with Id: {account.Id}");        
+    {       
 
         _context.Accounts.Add(account);
         await _context.SaveChangesAsync();
@@ -61,19 +68,22 @@ public class AccountsController : ControllerBase
     }
 
     public class bodyObject { public int userId { get; set; } public int accountId { get; set; } }
-    [HttpPost("joinUserAccount")]
+    [HttpPost("User")]
     public async Task<ActionResult> JoinUserAccount([FromBody] bodyObject request)
     {
+        if (!await AuthorizeUser(request.accountId)) return Forbid();
+
         var user = await _context.Users.FindAsync(request.userId);
         var account = await _context.Accounts.FindAsync(request.accountId);
 
         if (user == null) return NotFound($"User: {request.userId} not found");
         if (account == null) return NotFound($"Account: {request.accountId} not found");
 
-        var existingJoin = await _context.UserAccounts
-            .FirstOrDefaultAsync(ua => ua.UserId == request.userId && ua.AccountId == request.accountId);
 
-        if (existingJoin != null) return BadRequest($"User: {request.userId} is already connected to this account: {request.accountId}");
+        if (await BelongsToUser(request.accountId, request.userId))
+        {
+            return BadRequest($"User: {request.userId} is already connected to this account: {request.accountId}");
+        }
 
         var userAccount = new UserAccount
         {
@@ -84,16 +94,64 @@ public class AccountsController : ControllerBase
         _context.UserAccounts.Add(userAccount);
         await _context.SaveChangesAsync();
 
-        return Ok("User association successfully created");        
+        return Ok("User association successfully created");
+    }
+
+    [HttpDelete("User")]
+    public async Task<ActionResult> DeleteUserAccount([FromBody] bodyObject request)
+    {
+        if (!await AuthorizeUser(request.accountId)) return Forbid();
+
+        var userAccount = new UserAccount
+        {
+            UserId = request.userId,
+            AccountId = request.accountId
+        };
+
+        try
+        {
+            _context.UserAccounts.Remove(userAccount);
+            await _context.SaveChangesAsync();
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            if (!await BelongsToUser(request.accountId, request.userId))
+            {
+                return BadRequest($"Association doesn't exist! User: {request.userId} Account: {request.accountId}");
+            }
+            if (!AccountExists(request.accountId))
+            {
+                return NotFound($"No account found to update with Id: {request.accountId}");
+            }
+            if (!UserExists(request.userId))
+            {
+                return NotFound($"No user found to update with Id: {request.userId}");
+            }
+            else
+            {
+                throw;
+            }
+        }
+
+        return Ok("User association successfully deleted");
+        
+    }
+    
+    [HttpGet("User")]
+    public async Task<ActionResult<bool>> TestBelongs([FromBody] bodyObject request)
+    {
+        if (!await AuthorizeUser(request.accountId)) return Forbid();
+        return await BelongsToUser(request.accountId, request.userId);       
     }
 
 
-    [HttpPut("{Id}")]
-    public async Task<IActionResult> PutAccount(int Id, Account account)
+    [HttpPut()]
+    public async Task<IActionResult> PutAccount(Account account)
     {
-        if (Id != account.Id) return BadRequest($"Id: {Id} parameter doesn't equal account.Id: {account.Id}");        
 
-        if (!AccountExists(Id)) return NotFound($"No account found to update with Id: {account.Id}");
+        if (!await AuthorizeUser(account.Id)) return Forbid();            
+
+        if (!AccountExists(account.Id)) return NotFound($"No account found to update with Id: {account.Id}");
         
         _context.Entry(account).State = EntityState.Modified;
 
@@ -103,7 +161,7 @@ public class AccountsController : ControllerBase
         }
         catch (DbUpdateConcurrencyException)
         {
-            if (!AccountExists(Id))
+            if (!AccountExists(account.Id))
             {
                 return NotFound($"No account found to update with Id: {account.Id}");
             }
@@ -113,17 +171,21 @@ public class AccountsController : ControllerBase
             }
         }
 
-        return NoContent();
+        return Ok(account);
     }
 
     [HttpDelete("{Id}")]
     public async Task<IActionResult> DeleteAccount(int Id)
     {
+        if (!await AuthorizeUser(Id)) return Forbid(); 
+
         var account = await _context.Accounts.FindAsync(Id);
         if (account == null)
         {
             return NotFound($"No account found to delete with Id: {Id}");
         }
+
+        // _context.UserAccounts.RemoveRange(userAccounts); // ???
 
         _context.Accounts.Remove(account);
         await _context.SaveChangesAsync();
@@ -131,8 +193,35 @@ public class AccountsController : ControllerBase
         return NoContent();
     }
 
+    private bool UserExists(int Id)
+    {
+        return _context.Users.Any(e => e.Id == Id);
+    }
+
     private bool AccountExists(int Id)
     {
         return _context.Accounts.Any(e => e.Id == Id);
+    }
+    
+    private async Task<bool> BelongsToUser(int accountId, int userId)
+    {
+        var existingJoin = await _context.UserAccounts
+            .FirstOrDefaultAsync(ua => ua.UserId == userId && ua.AccountId == accountId);
+
+        if (existingJoin == null) return false;
+        else return true;
+    }
+
+    public async Task<bool> AuthorizeUser(int accountId)
+    {
+        if (User.IsInRole("Admin")) return true;
+
+        return await BelongsToUser(accountId, GetCurrentUserId());
+    }
+
+    private int GetCurrentUserId()
+    {
+        var userIdClaim = User.FindFirst(ClaimTypes.NameIdentifier)?.Value;
+        return int.Parse(userIdClaim ?? "0");
     }
 }
